requirements.txt
ultralytics
sahi
PyQt6
pyyaml
opencv-python